predict internals toggling
