+++ Content.Shared/_ST/Atmos/InternalsToggledEvent.cs
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: 2025 Janet Blackquill <uhhadd@gmail.com>
+//
+// SPDX-License-Identifier: LicenseRef-Wallening
+
+namespace Content.Shared._ST.Atmos;
+
+[ByRefEvent]
+public record struct InternalsToggledEvent(bool Online);
