+++ Content.Stellar.Client/Mobs/InternalsResourceBarSystem.cs
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: 2025 Janet Blackquill
+//
+// SPDX-License-Identifier: LicenseRef-Wallening
+
+using Content.Stellar.Shared.Mobs;
+
+namespace Content.Stellar.Client.Mobs;
+
+public sealed class InternalsResourceBarSystem : SharedInternalsResourceBarSystem;
